some changes about PDF

+++ ExTeX/src/java/de/dante/util/Unit.java
@@ -28,7 +28,7 @@ import de.dante.extex.interpreter.type.glue.GlueComponent;
  * This class implements a converter e.g for dimen values.
  *
  * @author <a href="mailto:m.g.n@gmx.de">Michael Niedermair</a>
- * @version $Revision: 1.9 $
+ * @version $Revision: 1.10 $
  */
 public final class Unit {
 
@@ -79,9 +79,9 @@ public final class Unit {
      * @param value the <code>Dimen</code>
      * @return Returns the <code>Dimen</code>-value as BP
      */
-    public static double getDimenAsBP(final Dimen value) {
+    public static float getDimenAsBP(final Dimen value) {
 
-        return ((double) value.getValue() * MULBP) / (DEN << SHIFT);
+        return ((float) value.getValue() * MULBP) / (DEN << SHIFT);
     }
 
     /**
